BEL-1377 Refactor out execution to its own file
